Переименование js файлов. Вынос функций для массивов в utils.js
