+++ js/main.js
@@ -1,1 +1,1 @@
-import './test-date.js';
+import './test-data.js';
